Update WME Closure Helper Beta.user.js
Fixing the bulk button closure button - unchecks the select all

Fixing the select all checkbox so if not all checkboxes are checked for the closures, they don't toggle on/off.  Ex: if there are 3 closures and the first and third are checked, if you hit select all the first and third will uncheck and the 2nd will check.